2006-1-1  Dan Williams  <dcbw@redhat.com>
	* src/nm-activation-request.c
		- Change dhcp_state member of the NMActRequest structure
			from guint8 to guint32

	* src/dhcp-manager/nm-dhcp-manager.[ch]
		- (nm_dhcp_manager_get_state_for_device): return guint32 rather
			than guint8 to match the dbus argument.  Turns out we were
			overwriting memory since we were passing in only a guint8


git-svn-id: http://svn-archive.gnome.org/svn/NetworkManager/trunk@1245 4912f4e0-d625-0410-9fb7-b9a5a253dbdc

+++ src/dhcp-manager/nm-dhcp-manager.c
@@ -138,12 +138,12 @@ void nm_dhcp_manager_dispose (NMDHCPManager *manager)
 }
 
 
-guint8 nm_dhcp_manager_get_state_for_device (NMDHCPManager *manager, NMDevice *dev)
+guint32 nm_dhcp_manager_get_state_for_device (NMDHCPManager *manager, NMDevice *dev)
 {
 	DBusMessage *	message;
 	DBusMessage *	reply;
 	char *		path;
-	guint8		state = 0;
+	guint32		state = 0;
 	DBusError		error;
 
 	g_return_val_if_fail (manager != NULL, 0);
